pull out some common functionality
