+++ src/open_gira/wind.py
@@ -1,12 +1,54 @@
 """
-Functions for reconstructing wind fields from track data.
+For reconstructing wind fields from track data.
 """
 
 import numpy as np
 import pyproj
+import xarray as xr
 
 import geopandas as gpd
 import pandas as pd
+
+
+# dimensions on which we have maximum wind speeds for
+WIND_COORDS: dict[str, type] = {
+    "event_id": str,
+    "latitude": float,
+    "longitude": float,
+}
+
+# Environmental pressure values in hPa / mbar (standard estimate of background
+# pressure away from the cyclone) are taken from the AIR hurricane model, table
+# 3 in Butke (2012).  Available at:
+# https://www.air-worldwide.com/publications/air-currents/2012/
+# the-pressures-on-increased-realism-in-tropical-cyclone-wind-speeds-through-attention-to-environmental-pressure/
+ENV_PRESSURE = {
+    "NI": 1006.5,
+    "SA": 1014.1,
+    "NA": 1014.1,
+    "EP": 1008.8,
+    "SI": 1010.6,
+    "SP": 1008.1,
+    "WP": 1008.3,
+}
+
+
+def empty_wind_da() -> xr.DataArray:
+    """
+    Return a maxium wind field dataarray with a schema but no data.
+
+    N.B. To concatenate xarray objects, they must all share the same
+    coordinate variables.
+    """
+    da = xr.DataArray(
+        data=np.full((0,) * len(WIND_COORDS), np.nan),
+        coords={
+            name: np.array([], dtype=dtype)
+            for name, dtype in WIND_COORDS.items()
+        },
+        name="max_wind_speed"
+    )
+    return da
 
 
 def power_law_scale_factors(z0: np.ndarray, z1: float, z2: float) -> np.ndarray:
